+++ policyengine_canada/variables/gov/cra/benefits/ccb/family_net_income.py
@@ -3,7 +3,7 @@ from policyengine_canada.model_api import *
 
 class family_net_income(Variable):
     value_type = float
-    entity = Person
+    entity = Household
     label = "Family net income"
     unit = CAD
     documentation = "Sum of head and spouse's line 23600 on tax return"